feat: 클래스 등록 성공 시 모달 닫기 추가
