+++ src/pages/Admin/hooks/useClass.ts
@@ -1,12 +1,9 @@
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { isAxiosError } from 'axios';
+import { useMutation, useQuery } from '@tanstack/react-query';
 
 import { ClassTableData } from '@/pages/Admin/types';
 import AdminClassApi from '@/services/admin/class';
 
 function useClass() {
-  const queryClient = useQueryClient();
-
   const { data, isLoading, isError } = useQuery<ClassTableData[]>({
     queryKey: ['admin-class'],
     queryFn: () => AdminClassApi.getClasses(),
@@ -18,18 +15,6 @@ function useClass() {
       type: string;
       description: string;
     }) => AdminClassApi.createClass(classData),
-    onSuccess: () => {
-      alert('클래스가 등록되었습니다.');
-      queryClient.invalidateQueries({ queryKey: ['admin-class'] });
-    },
-    onError: error => {
-      alert('클래스 등록에 실패했습니다.');
-      if (isAxiosError(error)) {
-        console.error(error.response?.data.result.message);
-        return;
-      }
-      console.error(error);
-    },
   });
 
   const subClassMutation = useMutation({
@@ -39,18 +24,6 @@ function useClass() {
       description: string;
       orderNumber: number;
     }) => AdminClassApi.createSubClass(subClassData),
-    onSuccess: () => {
-      alert('서브 클래스가 등록되었습니다.');
-      queryClient.invalidateQueries({ queryKey: ['admin-class'] });
-    },
-    onError: error => {
-      alert('서브 클래스 등록에 실패했습니다.');
-      if (isAxiosError(error)) {
-        console.error(error.response?.data.result.message);
-        return;
-      }
-      console.error(error);
-    },
   });
 
   return { data, isLoading, isError, classMutation, subClassMutation };
